関連する型が多いので、統合用のtypesを用意した
